requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
jupyter>=1.0.0
pyyaml>=6.0